DuckDbPySparkIOManager
new impl

lint
