+++ python_modules/libraries/dagster-duckdb-pandas/dagster_duckdb_pandas/duckdb_pandas_type_handler.py
@@ -4,13 +4,8 @@ import pandas as pd
 from dagster_duckdb.io_manager import DuckDbClient, _connect_duckdb, build_duckdb_io_manager
 
 from dagster import InputContext, MetadataValue, OutputContext, TableColumn, TableSchema
-from dagster._config.field_utils import config_dictionary_from_values
-from dagster._core.definitions.definition_config_schema import (
-    ConfiguredDefinitionConfigSchema,
-    convert_user_facing_definition_config_schema,
-)
+from dagster._core.definitions.configured_adapters import ConfiguredIOManagerAdapter
 from dagster._core.storage.db_io_manager import DbTypeHandler, TableSlice
-from dagster._core.storage.io_manager import IOManagerDefinition
 
 
 class DuckDBPandasTypeHandler(DbTypeHandler[pd.DataFrame]):
@@ -129,24 +124,6 @@ Examples:
             ...
 
 """
-
-
-class ConfiguredIOManagerAdapter(IOManagerDefinition):
-    def __init__(self, parent_io_manager, args):
-        ## TODO: coerce all strings to string source
-        super().__init__(
-            resource_fn=parent_io_manager.resource_fn,
-            config_schema=ConfiguredDefinitionConfigSchema(
-                parent_io_manager,
-                convert_user_facing_definition_config_schema(
-                    None
-                ),  # this is actually just replicating a bug that allows for too permissive of config
-                config_dictionary_from_values(
-                    args,
-                    parent_io_manager.config_schema.as_field(),
-                ),
-            ),
-        )
 
 
 class DuckDbPandasIOManager(ConfiguredIOManagerAdapter):
