+++ python_modules/libraries/dagster-duckdb-pyspark/dagster_duckdb_pyspark/__init__.py
@@ -1,2 +1,3 @@
 from .duckdb_pyspark_type_handler import DuckDBPySparkTypeHandler as DuckDBPySparkTypeHandler
+from .duckdb_pyspark_type_handler import DuckDbPySparkIOManager as DuckDbPySparkIOManager
 from .duckdb_pyspark_type_handler import duckdb_pyspark_io_manager as duckdb_pyspark_io_manager
