+++ python_modules/libraries/dagster-duckdb-pyspark/dagster_duckdb_pyspark/duckdb_pyspark_type_handler.py
@@ -1,8 +1,11 @@
+from typing import Optional
+
 import pyspark
 from dagster_duckdb.io_manager import DuckDbClient, _connect_duckdb, build_duckdb_io_manager
 from pyspark.sql import SparkSession
 
 from dagster import InputContext, MetadataValue, OutputContext, TableColumn, TableSchema
+from dagster._core.definitions.configured_adapters import ConfiguredIOManagerAdapter
 from dagster._core.storage.db_io_manager import DbTypeHandler, TableSlice
 
 
@@ -129,3 +132,11 @@ Examples:
             ...
 
 """
+
+
+class DuckDbPySparkIOManager(ConfiguredIOManagerAdapter):
+    def __init__(self, database: str, schema: Optional[str] = None):
+        super().__init__(
+            parent_io_manager=duckdb_pyspark_io_manager,
+            args={"database": database, "schema": schema},
+        )
